tests/requirements-test.txt
# 测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
fastapi>=0.109.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
openpyxl>=3.1.0
langchain>=1.0.0
langchain-community>=0.0.0
mock>=5.0.0
